META-205: Code readability improvement
VMS-13766: Fix code style
VMS-13766: Fix comments spelling
VMS-13766: Fixed
VMS-13766: Introduce connection timeout to avoid infinite UI loading state
VMS-13766: Monitor only logical state changes, unify state naming in comments
VMS-13766: Use std::chrono::milliseconds for timeout
VMS-13832: SAAS license keys support - code readability improvement
VMS-13894: Introduce error code for nx::update::Status
VMS-13894: Sort error codes by display priority

<==== Mercurial data ====>
Branch: vms_4.0

+++ vms/libs/common/src/nx/vms/common/p2p/downloader/file_information.h
@@ -60,9 +60,9 @@ public:
     QString absoluteDirectoryPath;
 
     // Calculates a progress for download, [0..100].
-    float calculateDownloadProgress() const;
+    int calculateDownloadProgress() const;
     // Calculates total number of bytes downloaded.
-    int calculateDownloadedBytes() const;
+    qint64 calculateDownloadedBytes() const;
 };
 
 #define FileInformation_Fields \
